fix: refactor http creation logic
