+++ src/main.ts
@@ -1,11 +1,10 @@
-import fs from 'fs';
-import https from 'https';
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { INestApplication, Logger, ValidationPipe } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
 import { ExceptionsFilter } from './utils/filters/exception.filter.util';
 import { ApiResponseUtil } from './utils/common/api.response.util';
+import { createHttpServer } from './utils/common/common.util';
 import { Environment } from './utils/utils.types';
 import { rateLimitMiddleware } from './utils/middlewares/rate-limit.middleware';
 import { RedisService } from './redis/redis.service';
@@ -42,14 +41,8 @@ async function bootstrap(): Promise<void> {
   });
 
   if (env === Environment.PRODUCTION) {
-    const httpsOptions = {
-      key: fs.readFileSync('path/to/private-key.pem'),
-      cert: fs.readFileSync('path/to/certificate.pem'),
-    };
-    const httpsServer = https.createServer(
-      httpsOptions,
-      app.getHttpAdapter().getInstance(),
-    );
+    const httpsServer = createHttpServer(app);
+
     httpsServer.listen(process.env.PORT, () => {
       Logger.log(`Application is running on: ${process.env.PORT}`);
     });
